+++ frontend/src/components/steps/GuestForm/GuestForm.tsx
@@ -1,4 +1,4 @@
-import CustomButton from '../ui/CustomButton/CustomButton'
+import CustomButton from '../../ui/CustomButton/CustomButton'
 import { UserRoundPlus, ArrowRight } from 'lucide-react'
 
 interface GuestFormProps {
